refactor : racingCar, racingCarModel => model 폴더로 이동
