+++ src/js/racingcar/model/racingCar.js
